refactor(precompiles): dedupe B20 golden test harness (BOP-449)
