+++ crates/common/precompiles/tests/b20_stablecoin_v1_golden.rs
@@ -19,45 +19,29 @@
 //! `BLESS_GOLDEN=1 cargo test -p base-common-precompiles --features test-utils \
 //!    --test b20_stablecoin_v1_golden -- --nocapture` and copy the printed `GOLDEN_ROOT` values.
 
-use alloy_primitives::{Address, B256, Bytes, LogData, U256, b256, keccak256};
+use alloy_primitives::{Address, B256, Bytes, U256, b256};
 use alloy_sol_types::{SolCall, SolError, SolEvent, SolValue};
 use base_common_genesis::BaseUpgrade;
 use base_common_precompiles::{
     B20_MAX_SUPPLY_CAP, B20PolicyType, B20StablecoinInit, B20StablecoinStorage, B20StablecoinToken,
     B20TokenRole, FakePolicyAccounting, IB20, IB20Stablecoin, NoopPrecompileCallObserver,
-    PermitArgs, PolicyVersion, Stablecoin, StablecoinV1, StablecoinVersion, StablecoinVersions,
+    PolicyVersion, Stablecoin, StablecoinV1, StablecoinVersion, StablecoinVersions,
     TokenAccounting,
 };
 use base_precompile_storage::{BasePrecompileError, HashMapStorageProvider, StorageCtx};
-use k256::ecdsa::SigningKey;
+
+mod common;
+use common::{
+    ADMIN, ALICE, BOB, CAROL, CHAIN_ID, MEMO, POLICY_ID, TOKEN, anvil_owner, bless_or_assert_gas,
+    bless_or_assert_root, hash_token_state, ok_true, signed_permit, u,
+};
 
 // --- fixtures ---------------------------------------------------------------
 
-const TOKEN: Address = Address::repeat_byte(0x22);
-const ADMIN: Address = Address::repeat_byte(0xAD);
-const ALICE: Address = Address::repeat_byte(0xA1);
-const BOB: Address = Address::repeat_byte(0xB0);
-const CAROL: Address = Address::repeat_byte(0xCA);
-const CHAIN_ID: u64 = 8453;
 const NAME: &str = "USD Coin";
 const SYMBOL: &str = "USDC";
 const CURRENCY: &str = "USD";
-const MEMO: B256 = B256::repeat_byte(0x77);
 const LOGIC: StablecoinV1 = StablecoinV1;
-
-/// A concrete (non-sentinel) ALLOWLIST policy id (type byte = 1, counter = 7).
-///
-/// Unconfigured scopes default to the `ALWAYS_ALLOW_ID` (0) EVM zero-slot, so
-/// blocking/executor guards must be exercised against an explicitly configured
-/// policy id like this one. Under V1 the top byte encodes the policy type, and
-/// authorization goes through membership: ALLOWLIST (type = 1) authorizes exactly
-/// its members, so `.allow(POLICY_ID, acct)` grants and unconfigured accounts are
-/// blocked. A BLOCKLIST id (top byte 0, e.g. a bare `7`) would invert every guard.
-const POLICY_ID: u64 = (1u64 << 56) | 7;
-
-// Anvil/Hardhat account 0 — well-known test key, never used in production.
-const PRIVATE_KEY: [u8; 32] =
-    alloy_primitives::hex!("ac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80");
 
 // --- pinned storage hashes (bless with BLESS_GOLDEN=1; see module docs) --------
 
@@ -119,16 +103,6 @@ const ROOT_GRANT_UNCHECKED: B256 =
     b256!("c83dd3df2a6f62c0775fb908d7a921f65c8e0e735bdb9b0adf7d1e489b657688");
 
 // --- harness ----------------------------------------------------------------
-
-/// `U256` from a small literal.
-fn u(n: u64) -> U256 {
-    U256::from(n)
-}
-
-/// The ABI encoding for a boolean-returning op (`transfer`/`approve`).
-fn ok_true() -> Bytes {
-    Bytes::from(true.abi_encode())
-}
 
 /// Fresh provider with an initialized `USD Coin` stablecoin at [`TOKEN`].
 fn fresh() -> HashMapStorageProvider {
@@ -204,40 +178,10 @@ fn last_topic0(storage: &HashMapStorageProvider) -> B256 {
     storage.get_events(TOKEN).last().expect("an emitted event").topics()[0]
 }
 
-/// Deterministic keccak hash of the per-case snapshot: the token's emitted events
-/// (topics + data) followed by its sorted `(address, slot, value)` storage triples.
-///
-/// A plain content hash (not an MPT state root). Events are included so a regression in
-/// an event's payload — indexed args or data not otherwise reflected in storage, e.g. a
-/// `Memo`'s bytes — is pinned here even though logs are not storage.
-fn hash_state(storage: HashMapStorageProvider) -> B256 {
-    let events: Vec<LogData> = storage.get_events(TOKEN).clone();
-    let mut triples: Vec<(Address, U256, U256)> = storage.into_storage().collect();
-    triples.sort();
-    let mut buf = Vec::with_capacity(triples.len() * 84 + events.len() * 64);
-    for log in &events {
-        for topic in log.topics() {
-            buf.extend_from_slice(topic.as_slice());
-        }
-        buf.extend_from_slice(&log.data);
-    }
-    for (addr, slot, value) in triples {
-        buf.extend_from_slice(addr.as_slice());
-        buf.extend_from_slice(&slot.to_be_bytes::<32>());
-        buf.extend_from_slice(&value.to_be_bytes::<32>());
-    }
-    keccak256(&buf)
-}
-
-/// Asserts the storage hash, or prints it under `BLESS_GOLDEN` for (re)pinning.
+/// Asserts the token's storage hash, or prints it under `BLESS_GOLDEN` for (re)pinning.
 #[track_caller]
 fn assert_root(label: &str, storage: HashMapStorageProvider, expected: B256) {
-    let got = hash_state(storage);
-    if std::env::var("BLESS_GOLDEN").ok().as_deref() == Some("1") {
-        println!("GOLDEN_ROOT {label} = {got:#x}");
-        return;
-    }
-    assert_eq!(got, expected, "V1 storage hash drift for `{label}`");
+    bless_or_assert_root(label, hash_token_state(storage, TOKEN), expected);
 }
 
 /// Grants `role` to `who` and bumps the role member count (setup only).
@@ -255,13 +199,6 @@ fn fund(token: &mut B20StablecoinStorage<'_>, who: Address, amount: U256) {
     token.set_total_supply(supply + amount).unwrap();
 }
 
-/// Recovers the anvil account-0 address from [`PRIVATE_KEY`].
-fn anvil_owner() -> Address {
-    let key = SigningKey::from_slice(&PRIVATE_KEY).unwrap();
-    let point = key.verifying_key().to_encoded_point(false);
-    Address::from_slice(&keccak256(&point.as_bytes()[1..])[12..])
-}
-
 /// The V1 EIP-712 domain separator for the token at [`TOKEN`] on [`CHAIN_ID`].
 fn domain_separator(storage: &mut HashMapStorageProvider) -> B256 {
     StorageCtx::enter(storage, |ctx| {
@@ -272,35 +209,6 @@ fn domain_separator(storage: &mut HashMapStorageProvider) -> B256 {
         );
         LOGIC.domain_separator(&token, CHAIN_ID).unwrap()
     })
-}
-
-/// Builds a validly-signed `permit` call for `owner`'s current nonce.
-fn signed_permit(
-    domain_sep: B256,
-    nonce: U256,
-    owner: Address,
-    spender: Address,
-    value: U256,
-    deadline: U256,
-) -> IB20::permitCall {
-    let mut args =
-        PermitArgs { owner, spender, value, deadline, v: 0, r: B256::ZERO, s: B256::ZERO };
-    let signing_hash = args.signing_hash(domain_sep, nonce);
-    let key = SigningKey::from_slice(&PRIVATE_KEY).unwrap();
-    let (sig, recid) = key.sign_prehash_recoverable(signing_hash.as_slice()).unwrap();
-    let bytes = sig.to_bytes();
-    args.r = B256::from_slice(&bytes[..32]);
-    args.s = B256::from_slice(&bytes[32..]);
-    args.v = if recid.is_y_odd() { 28 } else { 27 };
-    IB20::permitCall {
-        owner: args.owner,
-        spender: args.spender,
-        value: args.value,
-        deadline: args.deadline,
-        v: args.v,
-        r: args.r,
-        s: args.s,
-    }
 }
 
 // ============================================================================
@@ -2092,13 +2000,7 @@ fn golden_gas_footprints() {
         ("update_policy", (2, 1, 0)),
     ];
 
-    if std::env::var("BLESS_GOLDEN").ok().as_deref() == Some("1") {
-        for (label, counts) in &actual {
-            println!("GAS {label} = {counts:?}");
-        }
-        return;
-    }
-    assert_eq!(actual, expected, "storage-access footprint (sload, sstore, keccak256) drift");
+    bless_or_assert_gas(&actual, expected);
 }
 
 // ============================================================================
